Fix return type of fdo_blob_size (#95)
Previously, fdo_blob_size() was returning values of type int32_t, but
the return value was used as size_t in many places.

Since the typical return type of the size of a file is size_t, the
return type of fdo_blob_size() was changed to size_t.

This required following additional changes.
1. Update fdo_blob_size() to send 0 for error conditions.
2. The check on return value of fdo_blob_size() was updated to align
with this change.

+++ lib/m-string.c
@@ -54,7 +54,7 @@ static int read_fill_modelserial(void)
 	int ret = -1;
 	uint8_t def_serial_sz = 0;
 	uint8_t def_model_sz = 0;
-	int32_t fsize = 0;
+	size_t fsize = 0;
 
 	fsize = fdo_blob_size((const char *)SERIAL_FILE, FDO_SDK_RAW_DATA);
 	if ((fsize > 0) && (fsize <= MAX_DEV_SERIAL_SZ)) {
