+++ storage/include/storage_al.h
@@ -38,7 +38,7 @@ int32_t fdo_blob_read(const char *blob_name, fdo_sdk_blob_flags flags,
 int32_t fdo_blob_write(const char *blob_name, fdo_sdk_blob_flags flags,
 		       const uint8_t *buffer, uint32_t length);
 
-int32_t fdo_blob_size(const char *blob_name, fdo_sdk_blob_flags flags);
+size_t fdo_blob_size(const char *blob_name, fdo_sdk_blob_flags flags);
 
 int32_t create_hmac_normal_blob(void);
